remove unnceccesary files & refactor scss
