--- lib/components/Body.js
@@ -1,15 +1,0 @@
-const React = require('react');
-const ActionButton = require('./ActionButton');
-
-class Body extends React.Component {
-  render() {
-    return (
-      <section>
-        <h2>Location</h2>
-        <h2>Weather</h2>
-      </section>
-    );
-  }
-}
-
-module.exports = Body;
