+++ lib/index.jsx
@@ -1,3 +1,3 @@
-require('./new');
+require('./components/new');
 require('../CSS/reset.scss');
 require('../CSS/styles.scss');
